Extract document creation to a dedicated module
